Imgur handling removed from WebsiteHandler, put into own Handler
